Reorganize files, code, and comments
