+++ doc.go
@@ -1,0 +1,4 @@
+// Package iri contains facilities for working with Internationalized Resource Identifiers as specified in RFC 3987.
+//
+// RFC reference: https://www.ietf.org/rfc/rfc3987.html
+package iri
